+++ useragent_mock_test.go
@@ -3,7 +3,7 @@
 //
 // Generated by this command:
 //
-//	mockgen -destination=useragent_mocks_test.go -package=slackauth -source=useragent.go userAgentSetter
+//	mockgen -destination=useragent_mock_test.go -package=slackauth -source=useragent.go userAgentSetter
 //
 
 // Package slackauth is a generated GoMock package.
